requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==23.12.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
dep-logic==0.2.0
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.13.4
findpython==0.6.0
gunicorn==21.2.0
h11==0.14.0
hishel==0.0.26
httpcore==1.0.5
httpx==0.27.0
idna==3.7
installer==0.7.0
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
mypy-extensions==1.0.0
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pbs-installer==2024.4.1
pdm==2.14.0
platformdirs==4.2.2
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
pyproject_hooks==1.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
resolvelib==1.0.1
rich==13.7.1
ruff==0.0.281
shellingham==1.5.4
sniffio==1.3.1
socksio==1.0.0
starlette==0.37.2
tomlkit==0.12.4
tortoise-orm==0.20.1
truststore==0.8.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
unearth==0.15.1
uvicorn==0.30.1
virtualenv==20.25.1
zstandard==0.22.0

requests~=2.32.3
beautifulsoup4~=4.12.3
aiohttp~=3.10.5
aiofiles~=24.1.0